backend/requirements.txt
Django==5.1.*
djangorestframework==3.15.*
django-environ==0.11.*
django-cors-headers==4.4.*
psycopg[binary]==3.2.*
python-whois==0.9.4
beautifulsoup4==4.12.3
requests==2.32.3
tldextract==5.1.3
responses==0.25.3
